+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 27.0
 + **Breaking:** Remove `RvxElement` constructor options & default shadow root.
++ Fix: Unsupported methods `ReactiveMap` no longer silently fail.
++ Add `getOrInsert` & `getOrInsertComputed` support to `ReactiveMap`.
 + Add custom keys to `mapArray`, `<For>` and `forEach`.
 + Remove type definition bundles from npm package.